Fixed type bugs, added clawbot specific values

+++ include/project/chassis.h
@@ -12,8 +12,8 @@ class Chassis{
     void move_forward(float distance,int velocity);
     void point_turn(int angle,int velocity);
 
-    int turn_radius;
-    int wheel_circumference;
+    double turn_radius;
+    double wheel_circumference;
   protected:
     //Peripherals_t peripherals;
     int motor_speed;
